lambda/requirements.txt
boto3>=1.34.0
pyyaml>=6.0
python-dateutil>=2.8.2
